backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
cffi==2.0.0
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
Flask-SocketIO==5.6.1
gevent==26.4.0
gevent-websocket==0.10.1
greenlet==3.4.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
opencv-python==4.13.0.92
paramiko==3.5.0
pycparser==3.0
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
Werkzeug==3.1.8
wsproto==1.3.2
zope.event==6.1
zope.interface==8.3
pyserial==3.5
